Simplified BeaconBackportRewards contract
It is enough to have one 100% interval from May genesis to this month's
genesis and hardcode the number of groups available at the old beacon
contract at the moment of the new genesis. This way we can make it all
simpler. We also don't need excludeGroups property and exceess recipient
- no groups are excluded and we are going to allocated all rewards.

Improved unit tests to work on a data close to the real ones.