Make rewards upgradeable
A single rewards contract deployment is designed for only one keep factory.
Because keeps may be created simultaneously in different factories,
it would be quite complicated to accommodate different versions on one contract.
Instead, the rewards contract can be given the address of an upgraded version.
If the upgraded version is specified,
rewards tokens that would normally be returned to the unallocated pool
will instead be sent to the new rewards contract.
This will gracefully bleed out the old rewards contract's reward tokens
as the creation of new keeps (or beacon groups) dwindles over time.