Make sure the contract is properly funded before allowing for allocation
We had a problem with firstIntervalStart property - it had to be set in
a way that firstIntervalStart + termLength is in the future. Otherwise,
someone could call allocateRewards(0) before we fund the contract
properly and the allocation for the first interval would not be the same
as we'd want it to be.

Here we are making Rewards contract Ownable for only one purpose - the
owner has to mark the contract as properly funded calling `markAsFunded`
function. Only from then, rewards can be allocated.

Setting contract as funded does not stop from funding it even more in
future.