Do not decrease dispensed rewards on finalizing upgrade
We do not decrease `dispensedRewards` value when `finalizeUpgrade`
is called to make the contract semantics easier and more transparent.

Additionally, we improve invariants defined in this contract by adding
two that should always hold and clarifying the one already existing that
is guaranteed to hold only right after `receiveApproval` is called.

Starting with invariants, we have three:
- `totalRewards.sub(dispensedRewards) == unallocatedRewards` should
   always hold,
- `token.balanceOf(address(this)) >= totalRewards.sub(dispensedRewards)`
   should always hold,
- `token.balanceOf(address(this)) == totalRewards.sub(dispensedRewards)`
   should hold right after receiveApproval is called.

In regards to `finalizeUpgrade`, two options were considered:

Option 1. We do not decrease `totalRewards`, everything else stay as-is.
In this case:
- `totalRewards` represents all KEEPs that came into this contract and were
  processed by `receiveApproval`,
- `dispensedRewards` represents all KEEPs that have already left this contract,
  either by `receiveReward` or `finalizeUpgrade`.

Option 2. We do decrease `totalRewards`, we do not increase `dispensedRewards`.
In this case:
- `totalRewards` represents those KEEPs that came into this contract and were
  processed by `receiveApproval` and did not leave this contract by
  `finalizeUpgrade`.
- `dispensedRewards` represents all KEEPs that have already left this contract
  by `receiveReward`. It does not include those migrated to a new contract by
  `finalizeUpgrade`.

In addition to the invariants defined above, we added one more requirement:
after calling `finalizeUpgrade`, anyone can still call `receiveApproval` on the
old contract and this call should not fail but also produce valid end state of
the old rewards contract.

After considering various scenarios, it's been cocluded that both Option 1 and 2
produce consistent outcomes and all invariants holds however Option 2 offers
better semantics and that's what was done here.