Updated @dev docs for _getKeepAtIndex and _getCreationTime
Docs on those abstract functions stated that the implementation is
required to revert if the identifier is invalid or not in the expected
range. The implementation in BeaconBackportRewards and BeaconRewards
does not rever and after some investigation, it's been concluded that
additional checks are not really needed and we may save some gas on not
having them. Hence, the documentation update.

_getKeepAtIndex is used in:
* _findEndpoint
* _find

`_findEndpoint` starts the search with lower bound set to 0 and upper
bound set to `_getKeepCount() - 1`. It is guaranteed that values of
`lb` and `ub` with which `_getKeepAtIndex(lb)`/`_getKeepAtIndex(ub)`
are called must be in range `0 <= lb/ub < _getKeepCount()` so there is
no chance `_getKeepAtIndex` is called for a non-existing keep.

`_find` is used only by `_findEndpoint` so there is also no risk here.

_getCreationTime is used in:
* _findEndpoint
* _find
* _processKeep

`_findEndpoint` starts the search with lower bound set to 0 and upper
bound set to `_getKeepCount() - 1`. It is guaranteed that values of
`lb` and `ub` with which `_getKeepAtIndex(lb)`/`_getKeepAtIndex(ub)`
are called must be in range `0 <= lb/ub < _getKeepCount()` so there is
no chance `_getCreationTime` is called for a non-existing keep.

`_find` is used only by `_findEndpoint` so there is also no risk here.

`_processKeep` is called only from `receiveReward` and `reportTermination`.
Those two functions are public but they have `factoryMustRecognize`
modifier that ensures `keepIdentifier < _getKeepCount()` so there
is no chance `_getCreationTime` is called for a non-existing keep.