Add CurveRewardsEscrowBeneficiary and tests
Also fixes tests for the basic beneficiary behavior, since the
beneficiary is now always expected to be a contract with a
__escrowSentTokens handler.

CurveRewardsEscrowBeneficiary is a beneficiary contract that immediately
stakes any tokens received to a CurveRewards contract (or any contract
that can have `stake(uint256)` called on it with the token amount. The
tests involved ensure that if the target contract is a
CurveRewards-style contract, its internal behavior of taking the tokens
in question is executed.

A resolveAllLogs helper is pulled from tBTC, which should eventually be
moved to a shared test utils library for all Keep projects.