Update app structure
Adjust the app structure to the semantic hierarchy.
The app layout starts from the most important part- header and navigation. This
containers tell the user the most information about what they're
looking at and how the site hierarchy can be navigated.
The next layout item is a content section that renders the page content.
The last and least important layout item- footer is on the bottom of the
app hierarchy.