Configure MaxGasPrice and BalanceAlertThreshold with units
We wanted to configure values from MaxGasPrice and BalanceAlertThreshold
in a type that allow exceeding int64 range. This was required especially
for balance alerting as max int64 in wei is just around 9.22 ether.
We decided to parse the value with unit based on configuration in the
config file.
Here are the updates to the code that use latest keep-common.