Update resource tooltip component
Set a default styles for link to the resources page(+ arrow right) insted of a
link that looked like a button. Also the CTA should be more clear- `Go to
Resources` button sounds lik something will happen there where it's
really just a `learn more` type interaction. So here we also changed a
link text to `Learn more in Resources`.