Fix dapp build process via docker
Added a `COPY .env* /app/` in dashboard Dockerfile.

To get the current app version we added an `.env` that contains
`REACT_APP_VERSION` env variable. This env is provided by
`create-react-app` under the hood. For the `create-react-app` be able to
inject env variable the .env file is needed during the build process and
we missed the `COPY .env* /app/` in Dockerfile.