Refresh a dapp when an account has changed
This is a temporary solution to prevent a situation when a user changed
an account but data has not been updated. After migrate to redux the
dapp fetches data only once and updates data based on emitted events.
This solution doesn't support a case where a user changed an account. We
are going to address it in a follow up work.