Fetching ECDSA rewards
Added a function that fetches the available rewards based on the output
file from the merkle object generator which is stored in
`src/rewards/allocation/rewards.json` file for the given operators
between the interval 0 and the current interval. Note that rewards may
already be withdrawn so need to take into account `RewardsClaimed`
events from the `ECDSARewardsDistributor` contract.