BatchedPhasedEscrow
BatchedPhasedEsrow is a token holder contract allowing contract owner to
approve a set of beneficiaries of tokens held by the contract, to
appoint a separate drawee role, and allowing that drawee to withdraw
tokens to approved beneficiaries in phases.

This contract is not an ultimate solution for a secure escrow with a
separate role withdrawing tokens. Notably, it does not limit the amount
or frequency of withdrawals so compromised drawee can drain the escrow
funds to one of the selected approved beneficiaries. The escrow owner
should proceed with caution and do not transfer all funds to
BatchedPhasedEscrow.