Increase message handler throttle
In terms of DKG, current message handler throttle
is prepared to handle `4` retransmissions of
`64` messages in given DKG state. A standard DKG
messaging state lasts for `1` delay block and `5`
active blocks when a retransmission is made on
each block. However, the point validation state
is active for `10` blocks during which it performs
`10` retransmissions. In this state, there is
a big chance the handler throttle buffer overflows
and messages will be dropped in result. As
response, we increase that buffer size to `8 * 64`
which should be a satisfying value for all states.