Fix `getChainIdFromV` function
The `getChainIdFromV` function takes `v` param in `Uint8Array`. To be
able to parse this param to `int` we should call `v.toString("hex")` as
a first step otherwise, the `parseInt` function won't be able to parse
correctly to `int` and return `undefined`.
Also added a jsdoc for this function insted of a comment.