Ledger integration for chain id > 255
The comments that the ledger eth app returns the lower 2 bytes are no
longer valid. Currently the eth app returns correct `v` param for
chainID > 255 according to the EIP-155. So we don't need to `signedV &
0xff === ledgerSignedV`. Updating the `@ledgerhq/hw-app-eth` dependency
to the latest version (v5.38) did the trick. There are no effects on
mainnet or ropsten build of the KEEP token dashboard since the chain id
for these networks are 1 and 3.