Separate role for funding TokenGeyser contract
We separated the role that is expected to fund the TokenGeyser contract
with distribution tokens. Previously it was expected that only owner can
fund the contract, now we let the funder to be configured by the owner.

This solution is needed in our use case where we will be transferring
tokens from an escrow via beneficiary contract. So we expect the funder
to be set to escrow's beneficiary address that will transfer the funds
to TokenGeyser contract.