Fund KeepTokenGeyser contract via receiveApproval
Handled funding TokenGeyser contract (`lockTokens` function) with use of
`receiveApproval` which is a part of `approveAndCall` pattern we use for
transferring tokens out of an escrow to rewards contracts.

We expect the transfer to come from a beneficiary contract that should
be set in `TokenGeyser` contract as a funder.

A function to fund the rewards (distribution) tokens requires an
additional parameter `durationSec` that we expect to be constatnt in our
case, but we let the contract owner to reconfigure it via setter.