Update LiquidityRewards abstract class- dapp
To withdraw tokens from reward pool:
* LPRewards contract- the `exit` function should be called w/o any args,
* TokenGeyser contract- the `unstake` function should be called with
number of deposit token to unstake and empty calldata.

To stake tokens in rewards:
* LPRewards contract- the `stake` function should be called with amount
of tokens,
* TokenGeyser contract- shouldbe called the same funcion as above but
with additional arg- empty calldata.

The cases above motivate to add the mapping functions which return the
correct function name and args to trigger the stake/withdraw
transaction. This commit will provide this.