Docker image for building celo blockchain client for development
Setting up Celo client locally for development is similar to setting up
Ethereum client locally, except that we need to use specially prepared
genesis file which is more complex than one for Ethereum.

The process of setting up keystore and doing chain genesis is similar as
for Ethereum client. Not very complex but time consuming and cumbersome.

The Docker image implemented here should help with this process.

A developer needs to build the image:
  docker build -t celo-dev .

And can later start Celo blockchain locally with:
  docker run -it -v $(pwd)/data:/data -p 8545:8545 -p 8546:8546 celo-dev

$(pwd)/data directory is where the chain data will be stored. Since it
is a Docker volume, chain data survives container restarts.

In case the developer wants to start from a fresh chain, it is enough to
remove chain data:
  rm -rf data/celo
or even remove the entire data directory:
  rm -rf data
and start the container again:
  docker run -it -v $(pwd)/data:/data -p 8545:8545 -p 8546:8546 celo-dev

Keystore will be automatically populated and chain genesis executed.