Switched to Docker's build-publish-action for client's CI
We experienced some problems with `satackey/action-docker-layer-caching`
that caused workflow runners to get out of a disk space. This is related
to the way the action stores layer incrementaly along with the
previously cached ones. See: https://github.com/satackey/action-docker-layer-caching/issues/55

There is also a solution from Docker that we can use for building images
but also on later stage of RFC-18 implementation to publish images to
registires. See: https://github.com/marketplace/actions/build-and-push-docker-images

With the new solution we can also use caching and hopefully it won't
cause problems we had before.