Fix bug with fetching wrappedTokenBalance
After the Transfer event was emitted we would update the `wrappedTokenBalance`
which caused a bug when user wanted to add more lp tokens after withdraw. It was
cause becaused we updated the 'wrappedTokenBalance' twice - first we fetched it
on `Transfer` event and the we add the value to the current state value in
redux.