Unification of GH Action solidity workflows between keep-core and keep-ecdsa
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvment and make
the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. For keep-core part of
the workflow responsible for testing and linting of Solidity build
was already developed as a GitHub Actions workflow. This commit
introduces four changes to the workflow, to make the implementation
similiar to analogous workflow in keep-core.
Those changes are:
1. Added triggering of workflow on push and pull request events for
'rfc-18/*' branches
2. Updated versions of actions
3. Added step that builds solidity contracts
4. Name of 'test' step changed to 'build-and-test'