Support reinitializing root saga
In some cases to reduce number of request to node we fetch data only
once and update it based on emitted events. In this case, there is a
need to reset redux store and reinitialize the root saga when a user
changes account. This will require some changes regarding dispatching
actions that trigger data fetching- it will be addressed in a follow-up
commits.