Create GH Actions workflow for publish of Go client
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvment and make
the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit enhances the
workflow for building and testing Go with a step publishing the created
image to the Google Container Registry.
For now only keep-test configuration is implemented. Implementation for
other environments will be done at later stage of work on RFC-18.