Interface for sources of truth implementation
To find holders of tokens it's required to fetch data from multiple
contracts (sources). It's expected that each source will be implemented
separately and all of them will use similiar interface. This is a
proposed interface class for implementations.
It's expected that all sources will require `context` and `finalBlock`.
The class should implement `getTokenHoldingsAtFinalBlock` function to do
all the magic and return a map of holders and balances.