Addresses types resolver
Addresses that are owners of tokens can be externally owned accounts or contracts.
Some tokens may be locked in contracts like TokenStaking or TokenGrant, but
their actual owner may be another address. Truth of sources should find actual
owners of the tokens. This class is used to filter out any addresses that are
contracts handled by truth of sources implementations.