Move build&publish of Token Dashboard dApp to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvment and make
the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit creates a
GitHub Action workfow for building of Token Dashboard dApp and
publishing it in keep-test Google Container Registry.