Removed skipping of undelegatedAt
undelegatedAt is set when the delegation owner starts undelegation.
Still, tokens are still staked at this point. When stake is recovered
from TokenStaking contract, the amount is set to 0 but undelegatedAt is
not updated.
If the stake amount is non-zero, it means it is still in TokenStaking contract and we should include it.