Include TokenStakingEscrow in calculations for token grants
Tokens can be redelegated via TokenStakingEscrow to another operator. We
need to verify any operators for the given grantId that were staked via
TokenStakingEscrow for any slashed or seized tokens.

Tokens from a grant that went to TokenStakingEscrow can be withdrawn or
revoked via that contract. We need to include these amounts in
calculations.