Recover unclaimed tokens
If tokens are left unclaimed in the contract we let the owner to recover
them after some period of time.

The period is configurable on allocation. If set to zero the recovery
won't be allowed. The value should be set in seconds, these are added to
the timestamp of allocation. After the period passes the owner will be
able to recover tokens to an address provided on recovery function call.