Fix the problem with arithmeticOperation variable
There is a problem with `arithmeticOperation` variable in `subscriptions.js`
file. We define it in a while loop without setting a value and assign some value
later in the if statement. The problem is that this value persists in further
iterations of the while loop. I fixed it by assigning a null value to it from
the start.

Still, the question "Why it didn't work properly before?" remains. I tested
similar code in sandboxes online and also in newly created projects and it
worked as it suppose to work - `arithmeticOperation` was undefined in each new
iteration.  Here the variable keeps the value from previous iteration and
because of that we were adding/removing KEEP tokens to the redux store when we
shouldn't have.