Change names of secrets
Name of the secret storing private key for the account used for
migration of contracts on Ethereum has been changed, so that we would
have similar naming to that for Celo chain account.
Also names of the ETH_HOSTNAME, ETH_NETWORK_ID and TRUFFLE_NETWORK
secrets have been enhanced with `KEEP_TEST` / `KEEP_TEST_ETH` prefix.