Pin dependencies versions for dashboard on ropsten
The dashboard is build and published automatically for keep-test on
master merge. It's using versions provided in `package.json`. With this
commit we're hardcoding the versions of contracts deployed already to
ropsten and used in keep-test environment. Once a change to dashboard is
merged to master (with these versions) it's published to the keep-test
bucket and we can deploy it to keep-test environment.

In the future we want it to be more flexible and the latest dependencies
for the given environment should be used.