Move generation of docs to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and to
make the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit creates
the GH Actions workflow for generation of documentation and removes the
CircleCI jobs doing the same
The commit also fixes issue with non-parsable character in one of the
docs, which was causing incorrect generation of PDF file.
Temporarily artifacts of the workflow are published as job artifacts in
GitHub, but will be ultimately published to Google Cloud Storage.