MetaMask auto-connect
If a user has MetaMask extension installed and is still connected to the
dapp site we use the MetaMask extension as a default sign-in option. The
user can disconnect a site for a given account directly in the MetaMask
extention (Metamask doesn't provide a function to disconnect a MetaMask
via their API).