Sign contract address along with destination address
To prevent reusability of the signatures in other contracts that require
signature of a address we included the contract's address in the message
to sign to increase the uniqueness of the message.

In order to claim an airdrop, users have to sign a message with the destination address.
However: the signature is not somehow tied to the deployed TokenDistributor contract and
could be re-used, either in another instance of the same contract or in another contract that
uses a similar signature.

A user signs a message to claim the Keep distribution. An attacker notices the signature
and reuses it in a different contract that requires a signature of a address. The results can
be unexpected for the user that produces the signature, because they do not want for their
signature to be re-used in a different context.