Fix displaying the paid-back delegations
The paid-back delegations from `StakingPortBacker` contract should be
displayed as delegations from liquid tokens, not from a grant. So we
need to filterout paid back delegations while fetching delegations from
a grant. Otherwise, the delegations will be duplicated and be displayed
under `Wallet Tokens` and `Granted Tokens` tab on the `Delegations`
page.