Handle eplorer mode `differen address` error
Prints the error in the modal window.
Also `abortWalletConnection()` was executer when this error appeared so I wrote
the condition to not abort the connection on explorer mode. The reason for that
is that when there was the error the user was automatically logged off of
current explorer mode session.