Improve fetching contrarct deployment block
In the previous implementation, the contract deployment block was
fetched when dapp was connecting to the wallet. In this case, it took
too long to connect the dapp with a wallet. Here we added a
`getContractDeploymentBlockNumber` function which returns the block
number from a cache or fetches this data from a node. So the deployment
block will be fetched when needed.