Set proper paths to nav links in Explorer Mode
We create useWalletAddressFromUrl hook to get an address of the wallet from url.
If the address returned is not empty, then it means we are in explorer mode. We
use it when using nav link to check if the user is currently in the explorer
mode and to keep the address of the wallet in the url while navigating through
the page.

Note: the Overview and Resource page does not contain address of the wallet in
the url even in explorer mode. It is not needed since we are not displaying any
info about the address there.

This means that when user navigate from /<address>/liquidity to /overview page
and the tries to return to liquidity it will display him /liquidity page instead
of <address>/liquidity (because we didn't get any wallet address from the url
while on /overview page and thus we didn't 'inject' the address to the route).

We should create a condition for the pages without the wallet address that
checks if current connector is an Explorer Mode connector and if yes then
redirect from the current url to the explorer mode url (unless it's Overview or
Resources page).

The condition should probably be created in userExplorerModeConnect hook (I
added a TODO there)