Add Go linting to the client.yml workflow
As part of work on RFC-18 a job linting the Go code has been added to
`client.yml` workflow. Action for linting of Go code has been also added
(as separate repository) as part of this work.

Part of the code has been temporarily commented out, to speed up testing
of the change. Before merging the change to master, code needs to be
uncommented. Similarily, workflow triggering comditions has been changed
to allaow for the testing. Before merging the changes, original
conditions need to be restored.