Split job generating docs to multiple smaller jobs
This approach will make it possible to run some of the actions in
parallel and will be more consistent with implementation which we have
for docs generation in `tbtc` project.
Commit also removes the upload of `/img` folder - the files in that
folder are needed to generation of `relay-states.pdf`, but we don't need
to upload them.