Stop pushing contracts artifacts to GC Bucket
We're no longer pushing migrated contracts artifacts to GCP Bucket as
the artifacts are being published to NPM registry. In the repos using
artifacts of keep-core we will rework the code to pull the contracts
addresses from the NPM packages instead of the bucket. It gives us
flexibility to execute migrations for different environment in modules
in pararell. Additionally we will no longer overwrite contracts data
with the latest run of the workflow.