Resolve latest dependencies in different ways depending on the event_name
If Token Dashboard workflow gets invoked manually (workflow_dispatch
event), it will update the dependencies by checking upstream modules
versions and then running `npm install --save-exact...`. This will make
sure that the chain of workflows executed automatically as a result of
manual trigger of the upstream build will use the artifacts built by
the workflows in that chain.
However, when Token Dashboard is executed as a result of `push` or
`pull_request` event, the information about upstream builds is not
availible. For building the module in that case, we will update the
dependencies based on values configured in 'package.json', by running
`npm update...` command.
Note: We publish only the artifacts which were built in workflows
started by `workflow_dispath` event.