Compile contracts before publishing
`NPM` workflow will be used to publish the contracts needed for the
setup of development environment to NPM registry. Before the change we
were not compiling the contracts before publishing, which was a problem,
because this way not all the files needed by the downstream workflows
were being published. As a result of contracts compilation system will
generate `*.json` files needed by the later workflows. We will then
place them in the `artifacts` directory. By the `package.json`
configuration this directory is being picked by the publish step along
with the `contracts/**/*.sol` files.
Apart from this change, couple of small updates were introduced to `NPM`
workflow, mainly update to higher versions of used external actions.