Upload docs to bucket's root directory
We want to treat `keep-core` as our "root" project. Due to that we don't
want to place the generated docs in `keep-core` subfolder, instead we
want to upload directly to the bucket's root directory (with docs
previews landing in the branch-named subfolder).
Note: Docs generated for other projects like `keep-ecdsa` and `tbtc`
will be stored in project-named subfolders.