Modify jobs' execution conditions
Before the change `docs-pdf` and `docs-html` jobs would be executed for
all pull requests, even if there were no changes to the doc-related
files.
After the change, thanks to providing additional conditions in the `if`
config field for 'docs-html` and `docs-pdf` jobs, the jobs will run for
PRs and merges to master only if there were changes to the docs-related
files. The jobs will be executed even if the previous job
(`docs-solidity`) fails.
Note: alternative solution would be to set 'continue-on-error: true` for
`docs-solidity` job and remove `if` conditions from `docs-html` and
`docs-pdf` jobs, but that would not result in workflow failure in case
of failure of `docs-solidity` (making it possible to not notice the
failure after merge to master).