Replace the CircleCI build status badge with GHA badges
We no longer use CircleCI to test/build the `keep-core` code. We need to
replace the CircleCI build status badge in the projec'ts README with the
similiar badges for GitHub Actions builds.
Two badges have been introduced in place of the old one:
* Solidity build (showing status of the last `Solidity` workflow
executed on `master` branch and triggered by the push event),
* Go build (showing status of the last `Go` workflow executed on
`master` branch and triggered by the push event).

The badges are displayed using `shields.io` tool, which was already used
for the Discord badge.

Also a badge was added in the README of the Token Dashboard:
* Token Dashboard / Testnet (showing status of the last `Token Dashboard
/ Testnet` workflow on `master` branch and triggered by the push event).

This time standard GitHub badge was used, as `shields.io` tool currently
does not support workflows with slash character in their name (see
https://github.com/badges/shields/issues/4559).