Add a keep library
Added a KEEP library that creates necessary contracts used in the KEEP
Token Dashboard and provides useful services. Currently, only
CoveragePool service is provided. In the feature, we are going to
migrate services from the `*.services.js` files to the KEEP object. It
will help KEEP consistent API in the entire app.