Resove coverage-pools contracts
We are adding a new functionality of coverage pools to the Keep Network.
The contracts required for its support are stored in the
`coverage-pools` repo and are deployed to the npm registry thanks to the
`coverage-pools/.github/workflows/contracts.yml` workflow.
In this commit we are updating the workflow building Token Dashboard to
use those published `coverage-pools` contracts.