Update custom actions paths after moving actions to `ci`
We've moved custom actions used in the Keep/tBTC Continuous Integration
from separate repositories to `ci` repository (under `actions` folder).
We now need to update all the actions' invocations.

Note that temporarily we call the actions on the feature branch - this
is only for testing purposes and will be changed to `@v1` once
tested.