Update `dashboard-format` GH workflow
We need to keep the `ifs` on the `step` level, as we want the
`dashboard-detect-changes` job (but not it's steps) to execute for
`workflow_dispatch` and `push` events. If we had the `if` on the job
level, the job woul be skipped for this type of events, resulting in
skipping of jobs that have it configured as a dependency in the `needs`
field.